Copy edit the text
I standardized some of the language and punctuation, fixed a few English language idioms, and made various stylistic touch ups here and there. I also fixed some of the titling and added one more link to the resources section. These are minor, presentational adjustments as this is otherwise a competent and compelling document.